change parseInt to parseFloat

+++ BalanceHistory.js
@@ -9,10 +9,10 @@ export default function GetBalanceHistory(
   for (var i = 0; i < transactions.length; i++) {
     var tx = transactions[i];
     if (tx.toAddress === address) {
-      currentBalance -= parseInt(tx.amount);
+      currentBalance -= parseFloat(tx.amount);
       history.push(currentBalance);
     } else {
-      currentBalance += parseInt(tx.amount);
+      currentBalance += parseFloat(tx.amount);
       history.push(currentBalance);
     }
   }
